GNG-492: Refactor ComponentDetails into its own file
