+++ app/src/pages/Components/index.js
@@ -1,20 +1,10 @@
 import { useEffect, useState } from "react";
 import styled from "styled-components";
-import Button from "../../components/Button";
 import Header from "../../components/Header";
-import LoadSpinner from "../../components/LoadSpinner";
-import TextInput from "../../components/TextInput";
 import { componentService } from "../../_services/component.service";
 
-// CKEditor components
-import { CKEditor } from "@ckeditor/ckeditor5-react";
-import ClassicEditor from "@ckeditor/ckeditor5-editor-classic/src/classiceditor";
-import Bold from "@ckeditor/ckeditor5-basic-styles/src/bold";
-import Italic from "@ckeditor/ckeditor5-basic-styles/src/italic";
-import Link from "@ckeditor/ckeditor5-link/src/link";
-import Paragraph from "@ckeditor/ckeditor5-paragraph/src/paragraph";
-
 // Page components
+import ComponentDetails from "./ComponentsDetails";
 import ComponentsList from "./ComponentsList";
 
 const Page = styled.div`
@@ -40,32 +30,6 @@ const ContentContainer = styled.div`
     color: #a12622;
     padding: 15px;
   }
-`;
-
-const ComponentDetails = styled.div`
-  background-color: white;
-  flex-grow: 1;
-  margin: 16px;
-  min-width: 300px;
-
-  div.component-field {
-    margin: 8px 0;
-
-    label {
-      display: block;
-      font-size: 16px;
-      font-weight: 700;
-    }
-    input {
-      font-size: 18px;
-      width: 450px;
-    }
-  }
-`;
-
-const Controls = styled.div`
-  background-color: white;
-  margin-top: 16px;
 `;
 
 function Components() {
@@ -218,72 +182,18 @@ function Components() {
           selectedType={selectedType}
           setComponentId={setComponentId}
         />
-        <ComponentDetails>
-          {isLoadingComponent ? (
-            <LoadSpinner />
-          ) : (
-            <>
-              {componentTitle && (
-                <div className="component-field">
-                  <label htmlFor="component-title">Title: </label>
-                  <TextInput
-                    id="component-title"
-                    value={componentTitle}
-                    onChange={(e) => setComponentTitle(e.target.value)}
-                  />
-                </div>
-              )}
-              {componentTitle && (
-                <div className="component-field">
-                  <span id="component-id">
-                    <strong>ID:</strong> {componentId}
-                  </span>
-                </div>
-              )}
-              {componentIntro && (
-                <CKEditor
-                  id="editor-contact-us"
-                  editor={ClassicEditor}
-                  config={{
-                    plugins: [Bold, Italic, Link, Paragraph],
-                    toolbar: {
-                      items: ["bold", "italic", "link"],
-                    },
-                    language: "en",
-                  }}
-                  data={componentIntro}
-                  onReady={(editor) => {
-                    console.log("Component editor ready.", editor);
-                  }}
-                  onChange={(event, editor) => {
-                    const intro = editor.getData();
-                    console.log({ event, editor, intro });
-                    setComponentIntro(intro);
-                  }}
-                  onBlur={(event, editor) => {
-                    console.log("Blur.", editor);
-                  }}
-                  onFocus={(event, editor) => {
-                    console.log("Focus.", editor);
-                  }}
-                />
-              )}
-              {componentId && (
-                <Controls>
-                  <Button onClick={() => handleSave(componentId)}>
-                    {isSaving ? "Saving" : "Save"}
-                  </Button>
-                </Controls>
-              )}
-              {isErrorSaving && (
-                <p className="error">Could not save component changes.</p>
-              )}
-            </>
-          )}
-          {isErrorComponent && (
-            <p className="error">Could not fetch component details.</p>
-          )}
-        </ComponentDetails>
+        <ComponentDetails
+          componentId={componentId}
+          componentIntro={componentIntro}
+          componentTitle={componentTitle}
+          handleSave={handleSave}
+          isErrorComponent={isErrorComponent}
+          isErrorSaving={isErrorSaving}
+          isLoadingComponent={isLoadingComponent}
+          isSaving={isSaving}
+          setComponentIntro={setComponentIntro}
+          setComponentTitle={setComponentTitle}
+        />
       </ContentContainer>
     </Page>
   );
